Pull from master to User branch

+++ src/main/java/A/A.java
@@ -3,6 +3,11 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+
+// Test Pull request Branch permission 2nd:
+// Check for at least 1 approval from default reviewers > 
+// chnged to  >
+// Check for at least 1 approval
 package A;
 import AP3.AP3;
 import API.API;
@@ -808,7 +813,7 @@ public class A extends javax.swing.JFrame {
     
     public static double sleep = 0.5;
     public static double Timeout = 15;
-    public static double Wait = 3;
+    public static long Wait = 3;
 
     public static String t;
     public static String err;
